% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\title{Calculate fst between two populations in an ms output}
\usage{
fst(x, s1 = 24, s2 = 60)
}
\arguments{
\item{x}{the vector of alleles from ms}

\item{s1}{pop size of the first population (first to appear in ms output)}

\item{s2}{pop size of the second population (second to appear in ms output)}
}
\description{
This function allows you to load PFAM results to datafra,e
}
\examples{
fst(x)
}
\keyword{fst}
